requirements_full.txt
crawl4ai
supabase
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4

